fixing strict type errors 1

+++ src/app/select-davenfors/select-davenfors.component.ts
@@ -10,15 +10,15 @@ import { AdminService } from '../admin/admin.service';
 })
 export class SelectDavenforsComponent implements OnChanges {
 
-    @Input() category: string; //Need this for displaying empty-list
+    @Input() category: string=''; //Need this for displaying empty-list
 
     @Input()
-    davenfors: Davenfor[];
+    davenfors: Davenfor[] = [];
 
     @Input()
-    adminPermission: boolean; //Indicating if admin is logged in, sending to http customized url's
+    adminPermission: boolean=false; //Indicating if admin is logged in, sending to http customized url's
 
-    constructor(public guestService: GuestService, public adminService:AdminService) {
+    constructor(public guestService: GuestService, public adminService: AdminService) {
     }
 
     ngOnChanges(): void {
@@ -26,10 +26,15 @@ export class SelectDavenforsComponent implements OnChanges {
 
     onDelete(davenfor: Davenfor) {
         if (confirm(`Are you sure you want to delete the name ${davenfor.nameHebrew}?`)) {//Get user permission before proceeding
-            if (this.adminPermission) { //Sent from admin user
-                this.adminService.deleteDavenfor(davenfor.id, davenfor.nameEnglish);
+            if (davenfor.id != undefined && davenfor.nameEnglish != undefined) {
+                if (this.adminPermission) { //Sent from admin user
+                    this.adminService.deleteDavenfor(davenfor.id, davenfor.nameEnglish);
+                }
+                this.guestService.deleteDavenfor(davenfor.id, davenfor.nameEnglish);
             }
-            this.guestService.deleteDavenfor(davenfor.id, davenfor.nameEnglish);
+            else
+                console.error('Davenfor ID is ' + davenfor.id + ' and nameEnglish is ' + davenfor.nameEnglish);
+
         }
     }
-}
+}
